requirements.txt
click
gitpython
PyGithub
requests
